removed duplicated image tests from kg-lexical-html-renderer
