+++ packages/kg-lexical-html-renderer/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "c8": "7.12.0",
+    "jsdom": "^20.0.3",
     "mocha": "10.1.0",
     "should": "13.2.3",
     "sinon": "14.0.2"
